+++ packages/plugins/src/plugins/drift/constants.ts
@@ -17,9 +17,11 @@ export const perpMarketsIndexesKey = 'perp-markets-indexes';
 export const oracleToMintKey = 'oracle-to-mint';
 
 export const airdropUrl = 'https://airdrop.drift.trade/eligibility/';
+export const airdropS3Url = 'https://airdrop-3.drift.trade/eligibility/';
+export const distributorS3Pid = 'E7HtfkEMhmn9uwL7EFNydcXBWy5WCYN1vFmKKjipEH1x';
 export const driftDecimals = 6;
 
-export const airdropStatics: AirdropStatics = {
+export const airdropStaticsS1: AirdropStatics = {
   id: 'drift-airdrop-1',
   claimLink: 'https://drift.foundation',
   image: platformImage,
@@ -28,4 +30,15 @@ export const airdropStatics: AirdropStatics = {
   emitterLink: platformWebsite,
   claimStart: 1715860800000,
   claimEnd: 1723831200000,
+};
+
+export const airdropStaticsS3: AirdropStatics = {
+  id: 'drift-airdrop-3',
+  claimLink: 'https://drift.foundation/fuel-claim',
+  image: platformImage,
+  name: 'S3',
+  emitterName: platformName,
+  emitterLink: platformWebsite,
+  claimStart: 1751304600000,
+  claimEnd: 1763121600000,
 };