+++ packages/plugins/src/plugins/drift/types.ts
@@ -11,3 +11,20 @@ export type PreMarketPrice = {
 };
 
 export type PerpMarketIndexes = Array<[number, string]>;
+
+export type AirdropS3Response = {
+  start_amount: number;
+  end_amount: number;
+  unvested_amount: number;
+  claimed_amount: number;
+  unlocked_amount_claimed: number;
+  locked_amount_withdrawn: number;
+  locked_amount: number;
+  claimable_amount: number;
+  claimant: string;
+  merkle_tree: string;
+  mint: string;
+  error?: string;
+  end_ts: number;
+  start_ts: number;
+};