+++ packages/plugins/src/plugins/meteora/index.ts
@@ -12,6 +12,7 @@ import stakeForFeeFetcher from './stake2earn/stakeForFeeFetcher';
 import stakeForFeeVaultsJob from './stake2earn/stakeForFeeVaultsJob';
 import dammV2Job from './cpamm/cpammV2Job';
 import dlmmJob from './dlmm/dlmmJob';
+import pricingDbcJob from './dbc/pricingDbcJob';
 
 export const jobs: Job[] = [
   vaultsJob,
@@ -22,6 +23,7 @@ export const jobs: Job[] = [
   dlmmVaultsJob,
   stakeForFeeVaultsJob,
   dammV2Job,
+  pricingDbcJob,
 ];
 export const fetchers: Fetcher[] = [
   farmsFetcher,